Move configuration dto to be internal to publish config service
