+++ components/configuration/configuration-write-service/build.gradle
@@ -12,6 +12,7 @@ dependencies {
     implementation project(':libs:configuration:configuration-core')
     implementation project(':libs:lifecycle:lifecycle')
     implementation 'javax.persistence:javax.persistence-api'
+    implementation 'net.corda:corda-avro-schema'
     implementation 'net.corda:corda-base'
     implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
 }
