+++ components/configuration/configuration-write-service/src/main/kotlin/net/corda/configuration/write/publish/ConfigurationDto.kt
@@ -1,4 +1,4 @@
-package net.corda.libs.configuration.dto
+package net.corda.configuration.write.publish
 
 import net.corda.data.config.Configuration as ConfigurationAvro
