+++ libs/configuration/configuration-core/build.gradle
@@ -14,7 +14,6 @@ dependencies {
     implementation platform("net.corda:corda-api:$cordaApiVersion")
 
     implementation 'net.corda:corda-base'
-    implementation 'net.corda:corda-avro-schema'
     implementation "net.corda:corda-config-schema"
     implementation 'net.corda.kotlin:kotlin-stdlib-jdk8-osgi'
